API: Update Harga (In Progress)

+++ app/src/main/java/com/keerjain/crownstailor/data/entities/offer/OfferPrices.kt
@@ -2,8 +2,8 @@ package com.keerjain.crownstailor.data.entities.offer
 
 data class OfferPrices(
     val idPesanan: Int?,
-    val biayaJahit: Int?,
-    val biayaMaterial: Int?,
-    val biayaKirim: Int?,
-    val biayaJemput: Int?,
+    val biayaJahit: Float?,
+    val biayaMaterial: Float?,
+    val biayaKirim: Float?,
+    val biayaJemput: Float?,
 )
